revisitig this project. Doing some code improvements

+++ data/src/main/java/org/lucashos/data/db/dao/MovieDao.kt
@@ -8,7 +8,7 @@ import org.lucashos.data.domain.entity.MovieEntity
 @Dao
 interface MovieDao {
     @Query("select count(*) from favourite_movies where id = :movieId")
-    fun findFavourite(movieId: Int): Single<Int>
+    fun findFavourite(movieId: Long): Single<Int>
 
     @Insert(onConflict = OnConflictStrategy.IGNORE)
     fun insert(movie: MovieEntity): Completable
