+++ data/src/main/java/org/lucashos/data/domain/entity/MovieEntity.kt
@@ -4,4 +4,4 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "favourite_movies")
-class MovieEntity(@PrimaryKey val id: Int)
+class MovieEntity(@PrimaryKey val id: Long)
