+++ data/src/main/java/org/lucashos/data/service/TmdbApiService.kt
@@ -12,11 +12,11 @@ interface TmdbApiService {
     fun getTopMovies(@Query("page") page: Int = 1): Single<MoviesListResponse>
 
     @GET("movie/{id}")
-    fun getMovie(@Path("id") id: Int): Single<MovieDetailResponse>
+    fun getMovie(@Path("id") id: Long): Single<MovieDetailResponse>
 
 
     @GET("movie/{id}/similar")
-    fun getSimilarMovies(@Path("id") id: Int, @Query("page") page: Int = 1): Single<MoviesListResponse>
+    fun getSimilarMovies(@Path("id") id: Long, @Query("page") page: Int = 1): Single<MoviesListResponse>
 
     @GET("search/movie")
     fun searchMovies(@Query("query") title: String, @Query("page") page: Int): Single<MoviesListResponse>
